bpf: fix complexity issue on 4.19 kernel
Latest egress gateway changes introduced a complexity issue on 4.19
kernel.

This commit solves it by moving some of the egress gateway logic
(specifically the part responsible for encapsulating and redirecting
traffic to the egress gateway) to a different program:
EGRESS_GATEWAY_REDIRECT.

This in practice reduces the size of the IPV4_FROM_LXC one, allowing the
bpf_lxc program to be accepted by the verifier.

Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>