bpf: nodeport: move DSR detection from bpf_lxc into from-netdev
Let the nodeport code detect the DSR-specific Option/Extension, and create
a corresponding CT entry (with CT_EGRESS). This is similar to what
we do for svc traffic to a local backend.

Inside bpf_lxc we still flag the pod-level CT entry (same tuple, but
CT_INGRESS) as .dsr = true, and use that flag in the reply path to trigger
the RevDNAT.

Note that the CT tuple used for creating the SNAT entry has its L4 ports in
opposite order now. So we need to adjust the xlate_dsr() functions
accordingly.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>