bpf: nodeport: add RevDNAT stage for DSR in to-netdev
DSR replies currently get their RevDNAT in from-container. That doesn't
work for hostNetwork backends, or when from-container sends the packet off
to the stack for L7 policy.

We recently fixed the same scenario for local backends, by adding a
RevDNAT stage in to-netdev. Extend this code to also handle DSR replies.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>